requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=8.3.1
Flask>=2.0.1
tqdm>=4.62.0
matplotlib>=3.4.3
scikit-learn>=0.24.2
transformers>=4.30.0 